xvm: make exec.Code and exec.Context as interface (#526)
* xvm: make exec.Code and exec.Context as interface

Let exec.Code and exec.Context become interfaces,
so that we can reuse the resolver when we add the interpreter mode later.

+++ xvm/exec/wasm-rt.h
@@ -188,7 +188,7 @@ typedef struct {
                                 uint32_t param_len);
 
   /* 根据module和name返回对应的值 */
-  double (*wasm_rt_resolve_global)(void* context, char* module, char* name);
+  int64_t (*wasm_rt_resolve_global)(void* context, char* module, char* name);
 
 } wasm_rt_ops_t;
